FbConstraintCracCreator: move generated sources to xsd package
