+++ data/crac-creation/crac-creator-fb-constraint/pom.xml
@@ -42,6 +42,13 @@
                     </plugins>
                 </configuration>
             </plugin>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-javadoc-plugin</artifactId>
+                <configuration>
+                    <excludePackageNames>com.farao_community.farao.data.crac_creation.creator.fb_constraint.xsd.*</excludePackageNames>
+                </configuration>
+            </plugin>
         </plugins>
     </build>
