+++ data/crac-creation/crac-creator-fb-constraint/src/main/java/com/farao_community/farao/data/crac_creation/creator/fb_constraint/FbConstraint.java
@@ -6,6 +6,7 @@
  */
 package com.farao_community.farao.data.crac_creation.creator.fb_constraint;
 
+import com.farao_community.farao.data.crac_creation.creator.fb_constraint.xsd.FlowBasedConstraintDocument;
 import com.farao_community.farao.data.native_crac_api.NativeCrac;
 
 /**
