+++ data/crac-creation/crac-creator-fb-constraint/src/main/java/com/farao_community/farao/data/crac_creation/creator/fb_constraint/crac_creator/FbConstraintCracCreator.java
@@ -10,10 +10,10 @@ import com.farao_community.farao.data.crac_api.Crac;
 import com.farao_community.farao.data.crac_creation.creator.api.CracCreator;
 import com.farao_community.farao.data.crac_creation.creator.api.ImportStatus;
 import com.farao_community.farao.data.crac_creation.creator.api.parameters.CracCreationParameters;
-import com.farao_community.farao.data.crac_creation.creator.fb_constraint.CriticalBranchType;
+import com.farao_community.farao.data.crac_creation.creator.fb_constraint.xsd.CriticalBranchType;
 import com.farao_community.farao.data.crac_creation.creator.fb_constraint.FbConstraint;
-import com.farao_community.farao.data.crac_creation.creator.fb_constraint.FlowBasedConstraintDocument;
-import com.farao_community.farao.data.crac_creation.creator.fb_constraint.IndependantComplexVariant;
+import com.farao_community.farao.data.crac_creation.creator.fb_constraint.xsd.FlowBasedConstraintDocument;
+import com.farao_community.farao.data.crac_creation.creator.fb_constraint.xsd.IndependantComplexVariant;
 import com.farao_community.farao.data.crac_creation.util.ucte.UcteNetworkAnalyzer;
 import com.farao_community.farao.data.crac_creation.util.ucte.UcteNetworkAnalyzerProperties;
 import com.google.auto.service.AutoService;
