+++ data/crac-creation/crac-creator-fb-constraint/src/test/java/com/farao_community/farao/data/crac_creation/creator/fb_constraint/crac_creator/OutageReaderTest.java
@@ -8,7 +8,7 @@ package com.farao_community.farao.data.crac_creation.creator.fb_constraint.crac_
 
 import com.farao_community.farao.data.crac_api.Crac;
 import com.farao_community.farao.data.crac_api.CracFactory;
-import com.farao_community.farao.data.crac_creation.creator.fb_constraint.OutageType;
+import com.farao_community.farao.data.crac_creation.creator.fb_constraint.xsd.OutageType;
 import com.farao_community.farao.data.crac_creation.util.ucte.UcteNetworkAnalyzer;
 import com.farao_community.farao.data.crac_creation.util.ucte.UcteNetworkAnalyzerProperties;
 import com.powsybl.iidm.import_.Importers;
